Updated build script Added Package tests
