+++ bindings/xcframework/metadata_shortener.xcframework/ios-arm64-simulator/Headers/metadata-shortener.h
@@ -1,0 +1,6 @@
+// File automatically generated by swift-bridge.
+#include <stdint.h>
+struct __private__ResultPtrAndPtr __swift_bridge__$ext_generate_extrinsic_proof(void* call, void* signed_extras, void* additional_signed, void* metadata, uint32_t spec_version, void* spec_name, uint16_t base58_prefix, uint8_t decimals, void* token_symbol);
+struct __private__ResultPtrAndPtr __swift_bridge__$ext_generate_metadata_digest(void* metadata, uint32_t spec_version, void* spec_name, uint16_t base58_prefix, uint8_t decimals, void* token_symbol);
+
+
