--- bindings/xcframework/metadata_shortener.xcframework/ios-arm64_x86_64-simulator/Headers/module.modulemap
@@ -1,5 +1,0 @@
-module metadata_shortener {
-    header "SwiftBridgeCore.h"
-    header "metadata-shortener.h"
-    export *
-}
