+++ bindings/xcframework/metadata_shortener.xcframework/macos-arm64/Headers/module.modulemap
@@ -1,0 +1,5 @@
+module metadata_shortener {
+    header "SwiftBridgeCore.h"
+    header "metadata-shortener.h"
+    export *
+}
